+++ cookbooks/bookshop/recipes/ci.rb
@@ -1,3 +1,0 @@
-# We need the grails plugin.
-jenkins_cli 'install-plugin grails'
-jenkins_cli 'safe-restart'