+++ cookbooks/not-another-bookshop/metadata.rb
@@ -1,4 +1,4 @@
-name             'bookshop'
+name             'not-another-bookshop'
 maintainer       'YOUR_COMPANY_NAME'
 maintainer_email 'YOUR_EMAIL'
 license          'All rights reserved'